move schemas to config file
