+++ packages/admin/example.future.config.ts
@@ -1,16 +1,84 @@
-import { Config, createFramework, extractSchemaOptions } from 'core';
+import { createId } from '@paralleldrive/cuid2';
+import { Config, createFramework, extractSchemaOptions, IntegrationFieldTypeEnum } from 'core';
 import { GoogleIntegration } from 'future-google';
 import { MailchimpIntegration } from 'future-mailchimp';
 import { SlackIntegration } from 'future-slack';
+import { z } from 'zod';
 
-import {
-  BASE_RECORD_SCHEMA,
-  CREATE_NOTE_OUTPUT_SCHEMA,
-  CREATE_NOTE_SCHEMA,
-  CREATE_TASK_OUTPUT_SCHEMA,
-  CREATE_TASK_SCHEMA,
-  RECORD_SCHEMA,
-} from '@/domains/workflows/constants';
+enum Status {
+  ACTIVE = 'ACTIVE',
+  ARCHIVED = 'ARCHIVED',
+}
+
+const ObjectCategoryEnum = {
+  people: 'people',
+  companies: 'companies',
+  deals: 'deals',
+} as const;
+
+const CREATE_NOTE_SCHEMA = z.object({
+  title: z.string().trim().min(1, 'Required'),
+  description: z.string().optional().describe(`type::${IntegrationFieldTypeEnum.LONG_TEXT}`),
+  publicId: z
+    .string()
+    .optional()
+    .default(() => `kn_${createId()}`)
+    .transform(v => `kn_${createId()}`),
+});
+
+const CREATE_NOTE_OUTPUT_SCHEMA = z.object({
+  id: z.string(),
+  title: z.string(),
+  description: z.string().optional().describe(`type::${IntegrationFieldTypeEnum.LONG_TEXT}`),
+  publicId: z.string(),
+  status: z.nativeEnum(Status),
+});
+
+const CREATE_TASK_OUTPUT_SCHEMA = z.object({
+  id: z.string(),
+  name: z.string(),
+  description: z.string().describe(`type::${IntegrationFieldTypeEnum.LONG_TEXT}`),
+  isCompleted: z.boolean(),
+  status: z.nativeEnum(Status),
+  dueDate: z.string(),
+});
+
+const CREATE_TASK_SCHEMA = z.object({
+  name: z.string().trim().min(1, 'Required'),
+  description: z.string().optional().describe(`type::${IntegrationFieldTypeEnum.LONG_TEXT}`),
+  dueDate: z.string().datetime().optional(),
+});
+
+const BASE_RECORD_SCHEMA = z.object({
+  recordType: z.enum([ObjectCategoryEnum.people, ObjectCategoryEnum.companies, ObjectCategoryEnum.deals]),
+});
+
+const RECORD_SCHEMA = z.discriminatedUnion('recordType', [
+  z.object({
+    recordType: z.literal(ObjectCategoryEnum.companies),
+    data: z.object({
+      name: z.string().trim().min(1, 'Required'),
+    }),
+  }),
+  z.object({
+    recordType: z.literal(ObjectCategoryEnum.deals),
+    data: z.object({
+      name: z.string().trim().min(1, 'Required'),
+      amount: z.coerce.number(),
+      closeDate: z.string().datetime(),
+      pipeline: z.string().trim().min(1, 'Required'),
+      pipelineStage: z.string().trim().min(1, 'Required'),
+    }),
+  }),
+  z.object({
+    recordType: z.literal(ObjectCategoryEnum.people),
+    data: z.object({
+      firstName: z.string().trim().min(1, 'Required'),
+      lastName: z.string().trim().min(1, 'Required'),
+      email: z.string().email(),
+    }),
+  }),
+]);
 
 // // We have an admin db
 // // Enter secrets and shit it saves it to admin db for that integration
