+++ packages/admin/src/types/index.ts
@@ -1,29 +1,7 @@
 import { Integration } from '@/domains/plugins/types';
 import { AutomationBlueprintWithRelations } from '@/domains/workflows/types';
 
-import { IconName } from './icons';
-
 export interface Config {
   integrations: Integration[];
   workflows: AutomationBlueprintWithRelations[];
 }
-
-export const ObjectCategoryEnum = {
-  people: 'people',
-  companies: 'companies',
-  deals: 'deals',
-} as const;
-
-export const ObjectCategoryEnumArray = [
-  ObjectCategoryEnum.people,
-  ObjectCategoryEnum.companies,
-  ObjectCategoryEnum.deals,
-] as const;
-
-export type ObjectCategory = keyof typeof ObjectCategoryEnum;
-
-export const srtToIcon: Record<ObjectCategory, IconName> = {
-  people: 'people',
-  companies: 'companies',
-  deals: 'handshake',
-};
